texFields and labels stylings
