+++ client/src/pages/PlacingAnOrder/FillTheFromText/FillTheFromText.jsx
@@ -1,0 +1,13 @@
+import React from "react";
+import { Box, Typography } from "@mui/material";
+
+const FillTheFromText = () => {
+	return (
+		<Box sx={{ marginBottom: "36px" }}>
+			<Typography>Будь ласка, заповніть форму.</Typography>
+			<Typography>Оберіть спосіб оплати та спосіб доставки</Typography>
+		</Box>
+	);
+};
+
+export default FillTheFromText;
