+++ client/src/pages/PlacingAnOrder/PlacingAnOrder.scss
@@ -1,4 +1,9 @@
+// .category__title__wrapper {
+// 	margin: "0";
+// }
+
 .inputs-wrapper {
+	margin-top: 20px;
 	display: flex;
 	flex-direction: column;
 	gap: 30px;
