+++ client/src/pages/PlacingAnOrder/sxStyles/inputLabel.js
@@ -1,0 +1,5 @@
+export const inputLabel = {
+	marginBottom: "10px",
+	fontSize: "14px",
+	color: "rgba(0, 0, 0, 1)",
+};
